Conflicts:
	.env

+++ app/Http/Controllers/AccountController.php
@@ -8,6 +8,8 @@ use Hash;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Models\Account;
+use App\Models\Department;
+use App\Models\Area;
 use Helper, File, Session, Auth;
 
 class AccountController extends Controller
@@ -29,6 +31,9 @@ class AccountController extends Controller
     }
     public function index(Request $request)
     {          
+        if(Auth::user()->role != 3){
+            return redirect()->route('cost.index');
+        }
         $items = Account::where('role', '<', 3)->where('status', '>', 0)->orderBy('id')->get();        
         
         //$parentCate = Category::where('parent_id', 0)->where('type', 1)->orderBy('display_order')->get();
@@ -36,10 +41,14 @@ class AccountController extends Controller
         return view('account.index', compact('items'));
     }
     public function create()
-    {         
+    {        
+        if(Auth::user()->role != 3){
+            return redirect()->route('cost.index');
+        } 
         //$parentCate = Category::where('parent_id', 0)->where('type', 1)->orderBy('display_order')->get();
-
-        return view('account.create');
+        $departmentList = Department::all();
+        $areaList = Area::all();
+        return view('account.create', compact('departmentList', 'areaList'));
     }
     public function changePass(){
         return view('account.change-pass');   
@@ -72,11 +81,13 @@ class AccountController extends Controller
     }
     public function store(Request $request)
     {
-       
+        if(Auth::user()->role != 3){
+            return redirect()->route('cost.index');
+        }
         $dataArr = $request->all();
          
         $this->validate($request,[
-            'full_name' => 'required',
+            'name' => 'required',
             'email' => 'required|unique:users,email',
         ],
         [
@@ -95,7 +106,7 @@ class AccountController extends Controller
 
         $rs = Account::create($dataArr);
         if ( $rs->id > 0 ){
-            Mail::send('account.mail', ['full_name' => $request->full_name, 'password' => $tmpPassword, 'email' => $request->email], function ($message) use ($request) {
+            Mail::send('account.mail', ['name' => $request->name, 'password' => $tmpPassword, 'email' => $request->email], function ($message) use ($request) {
                 $message->from( config('mail.username'), config('mail.name'));
 
                 $message->to( $request->email, $request->full_name )->subject('Mật khẩu đăng nhập hệ thống');
@@ -108,6 +119,9 @@ class AccountController extends Controller
     }
     public function destroy($id)
     {
+        if(Auth::user()->role != 3){
+            return redirect()->route('cost.index');
+        }
         // delete
         $model = Account::find($id);
         $model->delete();
@@ -118,24 +132,29 @@ class AccountController extends Controller
     }
     public function edit($id)
     {
+        if(Auth::user()->role != 3){
+            return redirect()->route('cost.index');
+        }
         $detail = Account::find($id);
-        
-        return view('account.edit', compact( 'detail'));
+        $departmentList = Department::all();
+        $areaList = Area::all();
+        return view('account.edit', compact( 'detail', 'departmentList', 'areaList'));
     }
     public function update(Request $request)
     {
+        if(Auth::user()->role != 3){
+            return redirect()->route('cost.index');
+        }
         $dataArr = $request->all();
         
         $this->validate($request,[
-            'full_name' => 'required'            
+            'name' => 'required'            
         ],
         [
             'name.required' => 'Bạn chưa nhập họ tên'           
         ]);      
 
-        $model = Account::find($dataArr['id']);
-
-        $dataArr['updated_user'] = Auth::user()->id;
+        $model = Account::find($dataArr['id']);        
 
         $model->update($dataArr);
 
@@ -145,7 +164,9 @@ class AccountController extends Controller
     }
     public function updateStatus(Request $request)
     {       
-
+        if(Auth::user()->role != 3){
+            return redirect()->route('cost.index');
+        }
         $model = Account::find( $request->id );
